requirements.txt
# Stock AI Trading Framework Requirements
# Install with: pip install -r requirements.txt

# Data & Analysis
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
ta>=0.11.0

# Visualization
matplotlib>=3.7.0

# Logging
loguru>=0.7.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Broker API
alpaca-trade-api>=3.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# RL / Deep Learning
torch>=2.0.0

# News/RSS
feedparser>=6.0.10

# Deep RL (PyTorch)
torch>=2.2.0

